requirements.txt
Django==4.2.1
pytest==6.2.4
pytest-pythonpath==0.7.4
pytest-django==4.5.2
djangorestframework==3.14.0
Pillow==9.5.0
requests==2.30.0
